+++ packages/eslint-plugin/rules/__tests__/use-recommended-components.js
@@ -19,12 +19,13 @@ ruleTester.run( 'use-recommended-components', rule, {
 		"import * as UI from '@wordpress/ui';",
 
 		// Allowed @wordpress/ui components.
+		"import { Avatar } from '@wordpress/ui';",
 		"import { Badge } from '@wordpress/ui';",
 		"import { Icon } from '@wordpress/ui';",
 		"import { Link } from '@wordpress/ui';",
 		"import { Stack } from '@wordpress/ui';",
 		"import { Text } from '@wordpress/ui';",
-		"import { Badge, Icon, Link, Stack, Tabs, Text, Tooltip } from '@wordpress/ui';",
+		"import { Avatar, Badge, Icon, Link, Stack, Tabs, Text, Tooltip } from '@wordpress/ui';",
 
 		// Unlocked private APIs are only checked for denied names.
 		"import { privateApis } from '@wordpress/components'; import { unlock } from '../../lock-unlock'; const { SomethingElse } = unlock( privateApis );",