+++ packages/ui/src/avatar/index.ts
@@ -1,0 +1,5 @@
+import { Root } from './root';
+import { Image } from './image';
+import { Fallback } from './fallback';
+
+export { Root, Image, Fallback };